app/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
psycopg==3.2.6
psycopg-binary==3.2.6
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
